+++ src/aux-common/bots/BotEvents.ts
@@ -398,6 +398,7 @@ export interface ShellAction extends Action {
 export interface ShowToastAction extends Action {
     type: 'show_toast';
     message: string;
+    duration: number;
 }
 
 /**
@@ -915,10 +916,19 @@ export function pasteState(
  * Creates a new ShowToastAction.
  * @param message The message to show with the event.
  */
-export function toast(message: string): ShowToastAction {
+export function toast(message: string, duration?: number): ShowToastAction {
+    if (duration != null) {
+        return {
+            type: 'show_toast',
+            message: message,
+            duration: duration * 1000,
+        };
+    }
+
     return {
         type: 'show_toast',
         message: message,
+        duration: 2000,
     };
 }