+++ src/aux-server/aux-web/shared/SnackbarOptions.ts
@@ -4,6 +4,7 @@
 export default interface SnackbarOptions {
     visible: boolean;
     message: string;
+    duration?: number;
     action?: {
         type: string;
         label: string;