+++ CHANGELOG.md
@@ -30,6 +30,7 @@
     -   Added a `local` bot which is stored in the browser's local storage.
         -   The `local` bot is a bot that is unique to the device and channel.
         -   You can access the bot by querying for it: `getBot("#id", "local")`.
+    -   Added ability to set duration of toast, `plater.toast("message", durationNum)`.
 
 ## V0.10.3